Ah - glewinfo doesn't like the name mangling.
Only build glew if name mangling is off - that lets us get a run of glew
against libosmesa.

Looke like 2.0 is what is preset based on this reporting, so correct the
README.  Will have to see about the function signature complaints from Windows
and what can be done - looks like neither glew itself nor the 2.0 featurs will
work without the glprocs.h API.

If we're going to keep glprocs.h, we should get an example or two in place that
exercise it to make sure things are working.  Might be worth looking into the
Mesa 7.0.4 demos to see if there's anything there...