Explain why libosmesa is OpenGL 2.0 only
Digging in to the glewinfo output, figured out specifically why this
version of libosmesa is only reporting OpenGL 2.0 - although most needed
extensions are present, we only have support for GLSL v1.10. In light of
the extensive code churn in upstream history related to glsl handling,
the cost-benefit of supporting GLSL v1.20 to get full OpenGL 2.1 support
doesn't look worthwhile.

It would require analyzing the Mesa3D code history to find the most
mature code state of the slang code that could potentially work with the
rest of the libosmesa code base without drastic alterations.  Once that
was done and that code state was cleaned up any subsequent bug fixes
would need to be identified and backported.  Alternately, someone could
undertake new development starting with the current libosmesa state to
add the missing features.  That might be simpler (and would certainly be
less invasive) but I've not investigated which pieces are missing to know
how hard such an effort might be.

Someone would have to REALLY need GLSL 1.20 to make either of those
options worthwhile - in most scenarios I can imagine that need modern
features that badly, simply using the modern Mesa3D would make more
sense and be less effort.  The current BRL-CAD applications of libosmesa
are as a simple-to-build last resort way to guarantee scene displays and
source of offscreen rendering for regression testing.  Both of those use
only the fixed function OpenGL pipeline code - updating to GLSL v1.20
support offers no benefit. Barring outside interest and effort, we'll
stick with OpenGL v2.0 as the libosmesa max.