Default to O3, with a way to use O2
Using O3 makes things significantly faster - get a little over 300fps
with gears demo without it, and more than 400fps with it.  However,
for regression testing purposes O3 means less consistent behavior -
that's why we had default to O2.

Check for both flags now.  Default to faster rendering with an option to
disable it and use the slower O2 flag.