Do a little better about avoiding tcltest crashing
Just allocate one large-enough buffer up front rather than trying to
keep up with resizing it.  Also, use a mutex to be sure to not launch a
Dm_Render thread if one is already in progress  and only set up the font
context once, not on every DrawText.

Probably still doesn't fix everything, but seems to do better.